requirements.txt
# Web框架
Flask==2.3.3
flask-cors==4.0.0

# 数据库
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# LangGraph和AI相关（与Python 3.12兼容的版本）
langgraph==1.0.2
langchain==1.0.3
langchain-openai==1.0.2
langchain-community==0.4.1

# 环境配置
python-dotenv==1.0.0

# 数据处理
pandas==2.1.4
numpy==1.26.4

# 日期时间处理
python-dateutil==2.8.2

# 日志
loguru==0.7.2

# HTTP请求
requests==2.32.5

# 正则表达式增强
regex==2023.10.3

# JSON处理
ujson==5.8.0

# 类型检查
pydantic==2.12.3
